Refactored MemoryHandle
- Renamed PinnedMemoryHandle  to MemoryHandle.
- Dropped T parameter.
- Changed OwnedMemory<T> field to IKnown.
- Changed constructor to static Create<T> method (T required for
SizeOf(T) inside Add, and OwnedMemory<T> required to call
TryGetPointerInternal and TryGetArrayInternal).
