+++ src/System.Slices/System/Buffers/Memory.cs
@@ -62,7 +62,7 @@ namespace System
 
         public DisposableReservation<T> Reserve() => new DisposableReservation<T>(_owner);
 
-        public unsafe PinnedMemoryHandle<T> GetPinnedMemoryHandle() => new PinnedMemoryHandle<T>(_owner, _index);
+        public unsafe MemoryHandle GetPinnedMemoryHandle() => MemoryHandle.Create(_owner, _index);
 
         public unsafe bool TryGetPointer(out void* pointer)
         {
